requirements.txt
ultralytics
gradio
Pillow
numpy
